fabsf instead of fabs (float instead of double)

+++ float/float/float.c
@@ -571,7 +571,7 @@ static void do_rc_move(data *d)
 	else {
 		d->rc_counter = 0;
 		
-		if ((d->float_conf.remote_throttle_current_max > 0) && (d->current_time - d->disengage_timer > d->float_conf.remote_throttle_grace_period) && (fabs(d->throttle_val) > (double)0.02)) { // Throttle must be greater than 2% (Help mitigate lingering throttle)
+		if ((d->float_conf.remote_throttle_current_max > 0) && (d->current_time - d->disengage_timer > d->float_conf.remote_throttle_grace_period) && (fabsf(d->throttle_val) > 0.02)) { // Throttle must be greater than 2% (Help mitigate lingering throttle)
 			float servo_val = d->throttle_val;
 			servo_val *= (d->float_conf.inputtilt_invert_throttle ? -1.0 : 1.0);
 			d->rc_current = d->rc_current * 0.95 + (d->float_conf.remote_throttle_current_max * servo_val) * 0.05;
@@ -831,7 +831,7 @@ static void calculate_setpoint_target(data *d) {
 				}
 			}
 		}
-	} else if ((fabs(d->acceleration) > 15) &&					// this isn't normal, either wheelslip or wheel getting stuck
+	} else if ((fabsf(d->acceleration) > 15) &&					// this isn't normal, either wheelslip or wheel getting stuck
 			  (SIGN(d->acceleration) == SIGN(d->erpm)) &&		// we only act on wheelslip, not when the wheel gets stuck
 			  (d->abs_duty_cycle > 0.3) &&
 			  (d->abs_erpm > 2000))								// acceleration can jump a lot at very low speeds
